Extended FileUtils and using that in SnipSniper

+++ src/main/java/org/snipsniper/utils/FileUtils.java
@@ -10,6 +10,8 @@ public class FileUtils {
     }
 
     public static boolean delete(File file) {
+        if(!file.exists()) return true;
+
         if(!file.delete()) {
             LogManager.log("File (%c) could not be deleted!", LogLevel.WARNING, file);
             return false;
@@ -22,22 +24,41 @@ public class FileUtils {
     }
 
     public static boolean mkdir(File folder) {
+        if(folder.exists()) return true;
+
         if(!folder.mkdir()) {
-            LogManager.log("Folder (%c) could not be created.", LogLevel.WARNING, folder);
+            LogManager.log("Folder (%c) could not be created.", LogLevel.WARNING, folder.getAbsolutePath());
             return false;
         }
         return true;
     }
 
     public static boolean mkdirs(String file) {
-        return mkdir(new File(file));
+        return mkdirs(new File(file));
     }
 
     public static boolean mkdirs(File folder) {
+        if(folder.exists()) return true;
+
         if(!folder.mkdirs()) {
             LogManager.log("Folders (%c) could not be created.", LogLevel.WARNING, folder);
             return false;
         }
         return true;
+    }
+
+    public static boolean mkdirs(String... folders) {
+        File[] array = new File[folders.length];
+        for(int i = 0; i < folders.length; i++)
+            array[i] = new File(folders[i]);
+        return mkdirs(array);
+    }
+
+    public static boolean mkdirs(File... folders) {
+        boolean success = true;
+        for(File folder : folders)
+            if(!mkdirs(folder))
+                success = false;
+        return success;
     }
 }
